Use size_t instead of unsigned long.
size of unsigned long does not match size of itensor::Index for LLP64 systems like Windows, resulting in static_cast errors. Use size_t instead.

+++ itensor/detail/skip_iterator.h
@@ -34,7 +34,7 @@ class SkipIterator
     typedef typename __traits_type::reference reference;
     typedef typename __traits_type::pointer pointer;
 
-    typedef unsigned long size_type;
+    typedef size_t size_type;
 
     public:
